+++ src/streetworks/common/__init__.py
@@ -18,6 +18,7 @@ from .from_ban import from_ban
 from .from_bdtopo import from_bdtopo
 from .from_datavia import from_datavia
 from .from_datex2 import from_datex2
+from .from_gnaf import from_gnaf_address, from_gnaf_road
 from .from_jersey import from_jersey
 from .from_kartverket import from_kartverket
 from .from_linz import from_linz_address, from_linz_road, from_linz_road_section
@@ -111,4 +112,6 @@ __all__ = [
     "from_linz_address",
     "from_linz_road",
     "from_linz_road_section",
+    "from_gnaf_address",
+    "from_gnaf_road",
 ]